fix: Will no longer match and replace the first line it finds
I discovered this while setting up `pre-commit` hooks for `eslint` in another repo where I happened to list `eslint` early, with quotes.
Then `dependency-consistency` would replace the first instance of eslint correctly, but all plugins, which started with `eslint-` would be turned into `eslint@<version>-<rest of name>`.

The simple fix would be to change the order of dependencies, but it's better to fix it properly